fix(riasec): keep result page v2 wrapper within composer gate
